requirements.txt
clickhouse-connect==0.9.2
llama-index==0.14.5
llama-index-llms-openrouter==0.4.2
ruff==0.14.1
python-dotenv==1.2.1
beautifulsoup4==4.12.3
python-dotenv==1.2.1
pytest==9.0.1
requests==2.31.0
tqdm==4.67.1
pydantic==2.12.3
fastapi
uvicorn
sentence-transformers
numpy
